Update ATTRIB keyword to be more usable.
We now store where the ATTRIB is defined for reporting in errors.
The attributes now return hashes to be stored into the userdata
instead of needing to store it in a hashref passed as arguments.
Arguments to ATTRIBs are automatically split at commas and passed as
a list to the coderef.